fix: restore ci paths for local identity
